Added code comments. Refactored code, renamed classes.
