--- KingTech.P1Reader.NuGet/Message/FailureEvent.cs
@@ -1,8 +1,0 @@
-﻿namespace KingTech.P1Reader.Message;
-
-public class FailureEvent
-{
-    public DateTime? Timestamp { get; set; }
-    public long? RawTimestamp { get; set; }
-    public int? DurationInSeconds { get; set; }
-}
