--- KingTech.P1Reader.NuGet/SerialReader/ISerialReader.cs
@@ -1,7 +1,0 @@
-﻿namespace KingTech.P1Reader.SerialReader;
-
-internal interface ISerialReader
-{
-    bool Start();
-    bool Stop();
-}
