+++ KingTech.P1Reader/appsettings.Development.json
@@ -11,9 +11,9 @@
     "SerialPort": "DUMMY",
     "Format": "V502"
   },
-  "DsmrReaderPublisherSettings": {
-    "Host": "http://192.168.2.191:7777",
-    "ApiKey": "2Y6W3MVLCKZZ0HBWPRR26OKUQD1TY44RMBCDZ9ROT1J1CESCRO7IEJANNY6B2SS5",
-    "Timeout": "00:00:05"
-  }
+  //"DsmrReaderPublisherSettings": {
+  //  "Host": "http://192.168.2.191:7777",
+  //  "ApiKey": "2Y6W3MVLCKZZ0HBWPRR26OKUQD1TY44RMBCDZ9ROT1J1CESCRO7IEJANNY6B2SS5",
+  //  "Timeout": "00:00:05"
+  //}
 }
